+++ CHANGELOG.md
@@ -13,3 +13,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Hard-Login flow implementation
 - Fixed UserInfo json parser error
 - Copyright 
+- Possibility to transfer claims via the sdk interface